extracts logic into separate function
