+++ packages/next/src/pages/LivePreview/meta.ts
@@ -2,7 +2,7 @@ import type { Metadata } from 'next'
 
 import { getTranslation } from '@payloadcms/translations'
 
-import type { GenerateEditViewMetadata } from '../Document'
+import type { GenerateEditViewMetadata } from '../Document/getMetaBySegment'
 
 import { meta } from '../../utilities/meta'
